fix(dpp)!: Change have_voice to voice inside package
As mentioned in the same pull request, prefixes like have_ are discouraged in xmake packages.

BREAKING CHANGES: This WILL break xmake scripts that use the have_voice flag, IMO it's better to keep the previous have_voice flag as a deprecated alias of the new voice flag, and in the next version of D++ completely remove it from the package.