Added check that fcsnippets only run if fcld set
The test on the presence of gfortran may not guarantee the linker is
properly set.  The new test check that fcld is also set, for now to
gfortran since the test on the Fortran compiler only considers this
compiler.

Note: this way, fcsnippets are only run if --toolchain=gfortran is set.